Add git completion and git status to prompt
Reformat prompt to look like
daxelrod@hostname:~/dotfiles/home (branch *+) $
This diverges from the macOS default in that it moves the username to
before the hostname, uses the full path, and, of course, adds the git
status.

Add new function __source_git_extras to find the platform-specific
location of git-completion.bash and git-prompt.sh and source them if
available. Currently this function only supports macOS Xcode Tools
paths, but is intended to be extended.

Add new function __set_prompt to handle configuring the prompt. It is
smart enough to fall back to a gitless prompt if __source_git_extras
couldn't load the __git_ps1 function.

The Git status shows the branch name and the staged/unstaged dirty
status in color. I've encounterd a problem with this before where
Readline finds the wrong number of characters for the prompt, so I may
turn off color if that turns out to be the root cause.