Turn on mixed relative line numbers out of insert
Display relative line numbers, except for the line under the cursor,
which is displayed as an absolute line number. This should aid relative
motion without all that pesky counting.

When in insert mode or when not in focus, switch back to absolute line
numbers.

Add a custom color for the current line number so that the text isn't a
ghastly yellow.