Make tabs visible
Use a tab icon symbol to show where tabs are. (The other characters that
make up the tab are filled in with spaces).

This required finally configureing IndentGuidesOdd. We still want it to
have the default background, but it needs a subtle gray that matches
IndentGuidesEven for the foreground (which was previously never a
printable character.)